AUL#3871 LoginServiceProperties moved out of PropertiesLoginServiceProfileCredentialsProvider (being reused). License headers added.
