+++ src/main/scala/za/co/absa/loginsvc/athena/model/LoginServiceProperties.scala
@@ -13,22 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package za.co.absa.loginsvc.athena.ctx
+package za.co.absa.loginsvc.athena.model
 
-import java.util.Properties
-import java.util.concurrent.atomic.AtomicReference
-
-object DriverContext {
-
-  private val propsRef = new AtomicReference[Properties]()
-
-  def set(props: Properties): Unit = {
-    propsRef.set(props)
-  }
-
-  def get: Properties = {
-    propsRef.get()
-  }
-
-  // intentionally no clear method to keep the props for the lifetime of the JVM process
-}
+case class LoginServiceProperties(
+                                   user: String,
+                                   pass: String,
+                                   lsUrl: String,
+                                   jwt2tokenSvcUrl: String,
+                                   debug: Boolean = false
+                                 )
